\name{rgpu-package}
\alias{rgpu-package}
\alias{rgpu}
\docType{package}
\title{
(Undocumented) Functions for the Advanced C Course
}
\description{
\packageDescription{rgpu}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rgpu}
\packageIndices{rgpu}
** An overview of how to use the package, including the most important **
** functions **
}
\author{
\packageAuthor{rgpu}

Maintainer: \packageMaintainer{rgpu}
}
\references{
** Literature or other references for background information **
}
%** Optionally other standard keywords, one per line, from file KEYWORDS in **
%** the R documentation directory **
\keyword{ package }
\seealso{
** Optional links to other man pages, e.g. **
%** \code{\link[<pkg>:<pkg>-package]{<pkg>}} **
}
\examples{
%% ** simple examples of the most important functions **
plus(4, 5)
}
