\name{rpuDevice}
\alias{rpuDevice}
\title{Id of the GPU device under use by the active host thread}

\description{
     Query the Id of the GPU device on which the active host thread executes the device code.
}

\usage{
rpuDevice()
}

\value{
     Returns the Id of the GPU device on which the active host thread executes the device code.
}

\examples{
rpuDevice()
}

\keyword{misc}


